fixed error with missing user profile

+++ bookShare/views/remove_intrest.py
@@ -1,5 +1,3 @@
-from django.shortcuts import render
-from django.contrib.auth.models import User
 from bookShare.models import Intrest, UserProfile, Book
 from django.shortcuts import redirect
 from django.urls import reverse
@@ -7,9 +5,10 @@ from django.contrib.auth.decorators import login_required
 
 @login_required(login_url="bookShare:login")
 def remove_intrest(request,book_id):
-    book = Book.objects.get(book_id=book_id)
-    user_profile = UserProfile.objects.get(user = request.user)
-    if Intrest.objects.filter(book = book,user_profile = user_profile).exists():
-        Intrest.objects.get(book = book,user_profile = user_profile).delete()
+    if UserProfile.objects.filter(user = request.user).exists():
+        book = Book.objects.get(book_id=book_id)
+        user_profile = UserProfile.objects.get(user = request.user)
+        if Intrest.objects.filter(book = book,user_profile = user_profile).exists():
+            Intrest.objects.get(book = book,user_profile = user_profile).delete()
 
     return redirect(reverse('bookShare:book_info', kwargs={'book_id':book_id}))
